pyproject.toml
[tool.poetry]
name = "cdk-chalice-lite"
version = "0.1.6"
description = "Utility to facilitate deploying Chalice via the cdk"
authors = ["Stephan Fitzpatrick <knowsuchagency@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/knowsuchagency/cdk-chalice-lite"
homepage = "https://pypi.org/project/cdk-chalice-lite/"
keywords = [
    "cdk",
    "chalice",
    "productivity"
]


[tool.poetry.dependencies]
python = "^3.7"
"aws_cdk.core" = "^1.42.1"
"aws_cdk.aws_s3_assets" = "^1.42.1"
chalice = "^1.14.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
invoke = "^1.4.1"
toml = "^0.10.1"
pre-commit = "^2.4.0"
qypi = "^0.5.0"
black = "^19.10b0"
"aws_cdk.aws_sqs" = "^1.42.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
